+++ package.json
@@ -32,11 +32,11 @@
     "vue-eslint-parser": "^7.6.0"
   },
   "dependencies": {
-    "@curvefi/ethcall": "6.0.7",
     "axios": "^0.21.1",
     "bignumber.js": "^9.0.1",
+    "@curvefi/ethcall": "6.0.7",
     "ethers": "^6.11.0",
     "memoizee": "^0.4.15",
-    "web-worker": "^1.3.0"
+    "web-worker": "1.2.0"
   }
 }
